poetry, first try
